requirements.txt
dbt-postgres
apache-airflow-providers-postgres
apache-airflow-providers-http
dbt-postgres
requests
psycopg2-binary